Ascii Art Generator/requirements.txt.txt
flask
flask-cors
pillow
requests